Moved the cancelAllOpenOrders to PhotonController from PhotonPlugin Got rid of Photon.java in scripting
Modified the sorting behaviour in LatestMessageFunction - we now use the reference number that we assign 
to incoming messages instead of relying on the sending time. we may have to change this


git-svn-id: file:///space/source.marketcetera.org/root/trunk@6622 501a4a96-914b-0410-8454-f5988a4ca596

+++ source/trunk/plugins/org.marketcetera.photon.core/src/main/java/org/marketcetera/photon/core/OutgoingMessageHolder.java
@@ -9,7 +9,7 @@ public class OutgoingMessageHolder extends MessageHolder {
 	public OutgoingMessageHolder(Message message) {
 		super(message);
 	}
-	public OutgoingMessageHolder(Message message, int referenceNo) {
+	public OutgoingMessageHolder(Message message, long referenceNo) {
 		super(message, referenceNo);
 	}	
 }
